Automatically format Union queries
Union queries were originally excluded from the formatting because you *can* format them yourself, but it seems like the vast majority of use cases will benefit from the consistent automatic formatting. (The rare? exception would be those that format the Union queries themselves, and want to do it a different way than the formatter. It seems like this would be a rare edge case that we can address down the road if we need to.)