Restrict SQL keyword determination for Access SQL
Access SQL has a smaller set of reserved keywords than the combination of MSSQL, MySQL used in doctrine/sql-formatter. Since the formatter is mostly used with Microsoft Access, I added a second filter to determine whether the identified keyword is reserved in Microsoft Access. This allows you to use alias names that might be reserved words in other dialects as demonstrated in #487.