Load CodeModule before checking class instancing
In the testing database, we need to fully load the code module before we will see a correct value on the class instancing. (This would cause the initial unit testing to fail, but after opening the VBA IDE, all the tests would pass.) Now all the tests pass the first time when opening the database.