Prevent crash from options subform with no db open
If there is no database open and you open the options dialog. Microsoft Access will crash when the form is closed. This seems to be caused by the table data subform. We shouldn't really need to change options without a database open, so I am disabling this ability for now.