Don't verify the ribbon if not using it
When you install VCS, you have the option to install the ribbon. If you choose not to install it, then this class should not attempt to verify the ribbon when initializing. (This gives us the ability to disable the ribbon for development purposes, or if problems with the ribbon are causing Access to crash.)