Rework ribbon to support translation
Refactored the ribbon code to dynamically load the ribbon strings from a *.json file. This allows us to feed translated strings to the ribbon so the buttons can display in other languages. The nice thing about this approach is that all the translation is managed in the Access add-in, and translation is done through a single *.po file.
Also updated the TwinBASIC project to version BETA_622.