Reduce noise in query export from unneeded aliases
Depending on the state of the internal query optimizer, you may see alias columns come and go from the source files of the queries. This update (finally) sanitizes these unneeded attributes, resulting in a more consistent output. This has been on my wish list for years, so it is nice to finally have it done!  😄