Ignore error on initializing command bars class
In normal usage, there is no problem with the original implementation. However, if you build the add-in database from source multiple times in succession, the second call to CloseCurrentDatabase2 needed to close the database file when opened exclusively seems to (temporarily?) remove the template command bar object. This template is loaded into the add-in via the AfterBuild hook, so it seems to work fine to just ignore the error when the class initializes. @bclothier - Let me know if you see any concerns with ignoring this error. It seems to work fine in my testing, but you have a lot more experience in this area.  😄