Update source files to match component name case
When writing a file, the adSaveCreateOverWrite flag will not actually effect a change in the case of a file name. Added a comparison to delete the file if the case of the name has changed. (This would happen when renaming an object in Access. We want the source file name to follow the change.)