Normalize screen coordinates for form position
One of the ongoing issues people have encountered with this add-in is the noise introduced when forms are exported on systems with different monitor configurations. This update intends to normalize the saved screen coordinate positions to a single 1080p monitor at 96 dpi. The forms will automatically adjust when used on other size monitors, but this gives us a consistent output in the export files.