Normalize form designer viewport size
After more testing, it appears that what I thought was the cached form coordinates was actually the designer viewport area, which explains why these numbers would change when forms were saved when the Access application window was in different positions. We are now normalizing these values to a consistent output for the source files, and forcing a recompute when the forms are imported back into the database. This gives us consistent export files but doesn't break the layout of dialog forms which seem to rely on the saved viewport sizes. By saving each form in design view, we give Access the opportunity to recompute this geometry for the form object in the database.