Implement lEtErCaSiNg function from Mike Wolfe
After reviewing various approaches on handling the letter casing issues with VBA, I settled on a simple integrated approach that calls a function before export and after build to enforce the user's rules for letter casing as defined in a specially named class module. If this class does not exist in the current project, it does nothing. To use this built-in functionality, simply add a "clsStandardLetterCasing" class to your project as described in https://nolongerset.com/standardizelettercasing/ See #599